+++ ts-extensions.js
@@ -7,28 +7,32 @@
 
 module.exports = {
   rules: {
-['default-param-last']: 'off',
-['@typescript-eslint/default-param-last']: ["error"],
+    ['default-param-last']: 'off',
+    ['@typescript-eslint/default-param-last']: ['error'],
 
-['lines-between-class-members']: 'off',
-['@typescript-eslint/lines-between-class-members']: ["warn","always",{"exceptAfterSingleLine":true}],
+    ['lines-between-class-members']: 'off',
+    ['@typescript-eslint/lines-between-class-members']: [
+      'warn',
+      'always',
+      { exceptAfterSingleLine: true },
+    ],
 
-['no-dupe-class-members']: 'off',
-['@typescript-eslint/no-dupe-class-members']: ["error"],
+    ['no-dupe-class-members']: 'off',
+    ['@typescript-eslint/no-dupe-class-members']: ['error'],
 
-['no-loss-of-precision']: 'off',
-['@typescript-eslint/no-loss-of-precision']: ["error"],
+    ['no-loss-of-precision']: 'off',
+    ['@typescript-eslint/no-loss-of-precision']: ['error'],
 
-['no-redeclare']: 'off',
-['@typescript-eslint/no-redeclare']: ["error"],
+    ['no-redeclare']: 'off',
+    ['@typescript-eslint/no-redeclare']: ['error'],
 
-['no-throw-literal']: 'off',
-['@typescript-eslint/no-throw-literal']: ["warn"],
+    ['no-throw-literal']: 'off',
+    ['@typescript-eslint/no-throw-literal']: ['warn'],
 
-['no-unused-vars']: 'off',
-['@typescript-eslint/no-unused-vars']: ["warn",{"ignoreRestSiblings":true}],
+    ['no-unused-vars']: 'off',
+    ['@typescript-eslint/no-unused-vars']: ['warn', { ignoreRestSiblings: true }],
 
-['require-await']: 'off',
-['@typescript-eslint/require-await']: ["warn"],
-  }
-}
+    ['require-await']: 'off',
+    ['@typescript-eslint/require-await']: ['warn'],
+  },
+};